# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MemOS Server REST APIs",
    "description": "A REST API for managing multiple users with MemOS Server.",
    "version": "1.0.1"
  },
  "paths": {
    "/product/search": {
      "post": {
        "tags": [
          "Server API"
        ],
        "summary": "Search memories",
        "description": "Search memories for a specific user.\n\nThis endpoint uses the class-based SearchHandler for better code organization.",
        "operationId": "search_memories_product_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/APISearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/product/add": {
      "post": {
        "tags": [
          "Server API"
        ],
        "summary": "Add memories",
        "description": "Add memories for a specific user.\n\nThis endpoint uses the class-based AddHandler for better code organization.",
        "operationId": "add_memories_product_add_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/APIADDRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/product/create_cube": {
      "post": {
        "tags": [
          "Server API"
        ],
        "summary": "Create a new memory cube",
        "description": "Create a new memory cube for a user.\n\nMemory cubes are containers that store different types of memories (textual, activation, parametric).\nEach cube can be owned by a user and shared with other users.\n\n**Note on cube_id vs mem_cube_id:**\nThese terms are used interchangeably throughout the API:\n- `cube_id` is the canonical identifier for a cube\n- `mem_cube_id` appears in many legacy endpoints and means the same thing\n- When using other endpoints (search, add, chat), you can reference this cube using either term\n\n**Semantic Clarification:**\n- **Single mem_cube_id** (deprecated): Used in older endpoints to identify a single cube.\n  New code should use `readable_cube_ids` / `writable_cube_ids` lists instead.\n- **readable_cube_ids**: List of cube IDs the user can read from (used in search/chat)\n- **writable_cube_ids**: List of cube IDs the user can write to (used in add/chat)",
        "operationId": "create_cube_product_create_cube_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCubeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCubeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/product/register_cube": {
      "post": {
        "tags": [
          "Server API"
        ],
        "summary": "Register an existing memory cube",
        "description": "Register an existing memory cube with the MOS system.\n\nThis method loads and registers a memory cube from a file path or creates a new one\nif the path doesn't exist. The cube becomes available for memory operations.\n\n**Note on cube_id vs mem_cube_id:**\nThese terms are used interchangeably throughout the API. The registered cube can then\nbe referenced by its cube_id/mem_cube_id in other endpoints.\n\n**Current Status:**\nThis endpoint validates the registration request. Full registration functionality\nrequires architectural integration with MOSCore, which will be completed in a future update.",
        "operationId": "register_cube_product_register_cube_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterCubeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterCubeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/product/scheduler/status": {
      "get": {
        "tags": [
          "Server API"
        ],
        "summary": "Get scheduler running status",
        "description": "Get scheduler running status.",
        "operationId": "scheduler_status_product_scheduler_status_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "User ID",
              "title": "User Id"
            },
            "description": "User ID"
          },
          {
            "name": "task_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional Task ID to query a specific task",
              "title": "Task Id"
            },
            "description": "Optional Task ID to query a specific task"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/product/scheduler/task_queue_status": {
      "get": {
        "tags": [
          "Server API"
        ],
        "summary": "Get scheduler task queue status",
        "description": "Get scheduler task queue backlog/pending status for a user.",
        "operationId": "scheduler_task_queue_status_product_scheduler_task_queue_status_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "User ID whose queue status is requested",
              "title": "User Id"
            },
            "description": "User ID whose queue status is requested"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskQueueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/product/scheduler/wait": {
      "post": {
        "tags": [
          "Server API"
        ],
        "summary": "Wait until scheduler is idle for a specific user",
        "description": "Wait until scheduler is idle for a specific user.",
        "operationId": "scheduler_wait_product_scheduler_wait_post",
        "parameters": [
          {
            "name": "user_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Name"
            }
          },
          {
            "name": "timeout_seconds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 120.0,
              "title": "Timeout Seconds"
            }
          },
          {
            "name": "poll_interval",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 0.5,
              "title": "Poll Interval"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/product/scheduler/wait/stream": {
      "get": {
        "tags": [
          "Server API"
        ],
        "summary": "Stream scheduler progress for a user",
        "description": "Stream scheduler progress via Server-Sent Events (SSE).",
        "operationId": "scheduler_wait_stream_product_scheduler_wait_stream_get",
        "parameters": [
          {
            "name": "user_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Name"
            }
          },
          {
            "name": "timeout_seconds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 120.0,
              "title": "Timeout Seconds"
            }
          },
          {
            "name": "poll_interval",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 0.5,
              "title": "Poll Interval"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/product/chat/complete": {
      "post": {
        "tags": [
          "Server API"
        ],
        "summary": "Chat with MemOS (Complete Response)",
        "description": "Chat with MemOS for a specific user. Returns complete response (non-streaming).\n\nThis endpoint uses the class-based ChatHandler.",
        "operationId": "chat_complete_product_chat_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/APIChatCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/product/chat/stream": {
      "post": {
        "tags": [
          "Server API"
        ],
        "summary": "Chat with MemOS",
        "description": "Chat with MemOS for a specific user. Returns SSE stream.\n\nThis endpoint uses the class-based ChatHandler which internally\ncomposes SearchHandler and AddHandler for a clean architecture.",
        "operationId": "chat_stream_product_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateCubeResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Response status code",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Cube created successfully"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CreateCubeResponseData"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response data"
          }
        },
        "type": "object",
        "title": "CreateCubeResponse",
        "description": "Response model for cube creation."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SearchResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Response status code",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Response message"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response data"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "SearchResponse",
        "description": "Response model for search operations."
      },
...
    }
  }
}